+++ post-compile.sh
@@ -11,6 +11,7 @@ ln -sf /buildroot/bin/ambassador /buildroot/bin/watt
 ln -sf /buildroot/bin/ambassador /buildroot/bin/amb-sidecar
 ln -sf /buildroot/bin/ambassador /buildroot/bin/app-sidecar
 ln -sf /buildroot/bin/ambassador /buildroot/bin/aes-plugin-runner
+ln -sf /buildroot/bin/ambassador /buildroot/bin/traffic-manager
 
 # Also note there is a different ambassador binary, written in Python, that
 # shows up earlier in the shell PATH:
@@ -27,6 +28,7 @@ sudo ln -sf /opt/ambassador/bin/ambassador /opt/ambassador/bin/watt
 sudo ln -sf /opt/ambassador/bin/ambassador /opt/ambassador/bin/amb-sidecar
 sudo ln -sf /opt/ambassador/bin/ambassador /opt/ambassador/bin/app-sidecar
 sudo ln -sf /opt/ambassador/bin/ambassador /opt/ambassador/bin/aes-plugin-runner
+sudo ln -sf /opt/ambassador/bin/ambassador /opt/ambassador/bin/traffic-manager
 
 # entrypoint.sh, aes-plugin-runner, and the ABI stuff later in this file expect
 # these to be here